requirements.txt
langgraph>=0.2.36
httpx>=0.27.0
pytest>=8.0.0
pyyaml>=6.0
